src/coordinator/requirements.txt
setuptools
google-api-python-client
oauth2client
python-dateutil
pytz
jpype1
jaydebeapi
pandas
